WIP: using regular rust tests - fails because of multi-threading
There is a panic! as the event loop should be initialised once on the main thread however during rust standard test execution the tests are executed in parallel and are not run on the main thread.
The env_logger also panics as it is initialised multiple times.
It's better to keep these as "online tests" rather than standard rust tests.